# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Freelance CI API",
        "description": "API REST de la plateforme Freelance CI — connexion freelances, clients et entrepreneurs en Côte d'Ivoire. Paiements via Genius Pay (Mobile Money, Carte), authentification Supabase JWT.",
        "contact": {
            "name": "Freelance CI",
            "email": "contact@freelance-ci.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.6.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://api.freelance-ci.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
        "/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get admin dashboard stats",
                "operationId": "4faa4d17579606dd7357185b917f0c50",
                "responses": {
                    "200": {
                        "description": "Dashboard statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AdminDashboard"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List all users (admin)",
                "operationId": "0c3da1c25572511425f831ce3e86b849",
                "responses": {
                    "200": {
                        "description": "List of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/admin/users/{id}/status": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update user status (admin)",
                "operationId": "a7b0321c7740432dde2200e04c638bcd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "suspended",
                                            "banned"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User status updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to update user status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/admin/verifications": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List pending verifications (admin)",
                "operationId": "f678e7e65ea6fe90b97d15a9d639ca49",
                "responses": {
                    "200": {
                        "description": "List of pending verifications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Verification"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/admin/verifications/{id}/approve": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Approve a verification (admin)",
                "operationId": "c17dc43dcd6eae4b87ec66b8007ff7b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Verification ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verification approved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Verification"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to approve verification",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/admin/verifications/{id}/reject": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Reject a verification (admin)",
                "operationId": "f33cc481bdd95e43cb692dbfe8912f49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Verification ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Verification rejected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Verification"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to reject verification",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/admin/reports": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List reports (admin)",
                "operationId": "cd786d12a77981504ffe75ffe8c56d3f",
                "responses": {
                    "200": {
                        "description": "List of reports",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Report"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/admin/reports/{id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Resolve a report (admin)",
                "operationId": "9f967fc9aa796923206f2ff6f5d38d14",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Report ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "resolved",
                                            "dismissed"
                                        ]
                                    },
                                    "resolution_notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Report resolved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Report"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to resolve report",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AdminDashboard": {
                "properties": {
                    "total_users": {
                        "type": "integer"
                    },
                    "total_freelances": {
                        "type": "integer"
                    },
                    "total_clients": {
                        "type": "integer"
                    },
                    "total_projects": {
                        "type": "integer"
                    },
                    "total_contracts": {
                        "type": "integer"
                    },
                    "total_revenue": {
                        "type": "number",
                        "format": "float"
                    },
                    "pending_verifications": {
                        "type": "integer"
                    },
                    "open_disputes": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "description": "JWT Supabase — coller le token reçu après connexion",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "BearerToken": []
        }
    ]
}